% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{lk_ratio_grow}
\alias{lk_ratio_grow}
\title{Likelihood ratio for grow.}
\usage{
lk_ratio_grow(tree, current_node, sigma_2_y, sigma_2_mu)
}
\arguments{
\item{tree}{The current tree.}

\item{current_node}{The current pruned node.}

\item{sigma_2_y}{The current value of sigma^2 for y.}

\item{sigma_2_mu}{The current valur of sigma^2_mu.}
}
\value{
The likelihood ratio.
}
\description{
Likelihood ratio of the candidate tree and the previous 
tree, given that the action step was a grow.
}
\details{
The likelihood ratio for growth needs the 
joint distribution of each region (of each node), given the 
parameter sigma^2 (here, mu is considered to be integrated out)
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
